requirements.txt
asgiref==3.5.0
distlib==0.3.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.6.0
gitignore==0.0.8
gunicorn==20.1.0
Markdown==3.3.6
platformdirs==2.5.1
psycopg2==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
virtualenv==20.13.3
whitenoise==6.0.0
